pyproject.toml
[tool.poetry]
name = "multitask-perception"
version = "0.1.0"
description = "Unified multitask perception for autonomous driving: Detection, Segmentation, and Depth"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
license = "MIT"
packages = [{include = "multitask_perception", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"

# Core ML frameworks
torch = "^2.2.0"
torchvision = "^0.17.0"
timm = ">0.9.12"

# Computer Vision
opencv-python = "^4.9.0"
pillow = "^10.2.0"
albumentations = "^1.3.1"
scikit-image = "^0.22.0"

# Dataset handling
pycocotools = "^2.0.7"
scipy = "^1.12.0"

# Configuration
pyyaml = "^6.0.1"
yacs = "^0.1.8"

# Training utilities
tensorboard = "^2.15.0"
wandb = "^0.16.2"
tqdm = "^4.66.1"

# Utilities
numpy = "^1.26.3"
pandas = "^2.1.4"
matplotlib = "^3.8.2"
seaborn = "^0.13.1"

# CLI and UI
rich = "^13.7.0"
typer = "^0.9.0"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"

# Linting and formatting
black = "^23.12.1"
flake8 = "^7.0.0"
isort = "^5.13.2"
mypy = "^1.8.0"

# Jupyter
jupyter = "^1.0.0"
ipython = "^8.19.0"
ipykernel = "^6.28.0"

[tool.poetry.scripts]
train = "multitask_perception.tools.train:main"
eval = "multitask_perception.tools.eval:main"
infer = "multitask_perception.tools.infer:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/multitask_perception --cov-report=html --cov-report=term"
